requirements.txt
amqp==2.5.2
attrs==19.3.0
autopep8==1.5.2
backcall==0.1.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==2.2.13
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-silk==4.0.1
djangorestframework==3.11.0
fake-useragent==0.1.11
flake8==3.8.1
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==3.1.4
gprof2dot==2019.11.30
gunicorn==20.0.4
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
kombu==4.6.8
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.2
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.1
requests==2.23.0
sentry-sdk==0.15.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
uWSGI==2.0.19.1
vine==1.3.0
wcwidth==0.1.9
XlsxWriter==1.2.9
